% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.r
\docType{methods}
\name{annotate.gwas}
\alias{annotate.gwas}
\alias{annotate.gwas,GWAS,FeatureList-method}
\title{Annotate GWAS makers with genomic features}
\usage{
annotate.gwas(object, feature.list)

\S4method{annotate.gwas}{GWAS,FeatureList}(object, feature.list)
}
\arguments{
\item{object}{\code{\link{GWAS}} object}

\item{feature.list}{\code{FeatureList} object created with
\code{\link{hub.features}} or \code{\link{local.features}}}
}
\value{
\code{\link{AnnotatedGWAS}} object
}
\description{
Annotate GWAS makers with genomic features
}
\examples{
\dontrun{
query <- list(DNaseI = c("Dnasen", "broadPeak", "rep1"))
feature.list <- hub.features(query)
scz <- annotate.gwas(scz, feature.list)
}
}

