% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{as.GWAS,GRanges-method}
\alias{as.GWAS,GRanges-method}
\title{Create a GWAS object}
\usage{
\S4method{as.GWAS}{GRanges}(object, genome, marker, pvalue, zscore, or, beta)
}
\arguments{
  \item{object}{a \code{\link{data.frame}} or
  \code{\link[GenomicRanges]{GRanges}} object containing
  the required fields necessary to construct a
  \code{\link{GWAS}} object}

  \item{genome}{genome abbreviation used by UCSC (e.g.,
  hg19)}

  \item{marker}{name of the column in \code{object} that
  contains the marker (or SNP) identifiers}

  \item{pvalue}{name of the column in \code{object} that
  contains the GWAS association p-value for each marker}

  \item{zscore}{name of the column in \code{object} that
  contains the GWAS z-scores for each marker}

  \item{or}{name of the column in \code{object} that
  contains the GWAS odds ratio for each marker}

  \item{beta}{name of the column in \code{object} that
  contains the GWAS beta values (or regression
  coefficients) for each marker}
}
\description{
Create a \code{\link{GWAS}} object from a \link[GenomicRanges]{GRanges}
object
}

