\name{cache.features}
\alias{cache.features}
\title{Download uncached features}
\usage{
cache.features(feature.list, path)
}
\arguments{
  \item{feature.list}{\code{FeatureList} object created
  with \code{\link{hub.features}} or
  \code{\link{local.features}}}

  \item{path}{Path of directory containing features}
}
\value{
\code{FeatureList}
}
\description{
Download uncached features
}

