% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureList-methods.r
\name{Cached}
\alias{Cached}
\title{Access Cached column from FeatureList object}
\usage{
Cached(object)
}
\arguments{
\item{object}{\code{FeatureList} object}
}
\description{
Access Cached column from FeatureList object
}

