\name{calc.enrich}
\alias{calc.enrich}
\title{Calculate enrichment of all annotated features across thresholds}
\usage{
calc.enrich(object, stat, thresh.levels)
}
\arguments{
  \item{gwas}{GWAS GRanges object}

  \item{stat}{numeric vector containing statistic to which
  thresholds will be applied for each enrichment
  calculation}
}
\description{
Calculate enrichment of all annotated features across
thresholds
}

