% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotatedGWAS-methods.r
\name{consolidate}
\alias{consolidate}
\title{Consolidate each group of annotated features}
\usage{
consolidate(object)
}
\arguments{
\item{object}{\code{AnnotatedGWAS} object}
}
\value{
\code{AnnotatedGWAS} object
}
\description{
consolidate combines the features within each group to produce a new single
feature that comprises all of the original individual features
}

