% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotatedGWAS-methods.r
\name{fcols}
\alias{fcols}
\title{Extract features from AnnotatedGWAS object}
\usage{
fcols(object)
}
\arguments{
\item{object}{\code{AnnotatedGWAS} object}
}
\description{
Extract features from AnnotatedGWAS object
}

