% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features.r
\name{feature.labels}
\alias{feature.labels}
\title{Create pretty feature labels from the full RDataPaths}
\usage{
feature.labels(x)
}
\arguments{
\item{x}{characer vector of full RDataPaths}
}
\description{
Basically just strips off the extension from the basename
}

