\name{feature.search}
\alias{feature.search}
\title{Search AnnotationHub for features}
\usage{
feature.search(data.filter, genome, md, online = NULL)
}
\arguments{
  \item{data.filter}{named list of vectors}

  \item{online}{logical, should metadata be retrieved
  online}
}
\description{
Search AnnotationHub for features
}

