% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AnnotatedGWAS-methods.r
\name{featureIndex}
\alias{featureIndex}
\title{Access AnnotatedGWAS feature index}
\usage{
featureIndex(object)
}
\arguments{
\item{object}{\code{AnnotatedGWAS} object}
}
\description{
Access AnnotatedGWAS feature index
}

