% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FeatureList-class.r, R/FeatureList-methods.r
\docType{methods}
\name{featurelist}
\alias{Cached}
\alias{Cached,FeatureList-method}
\alias{FeatureList}
\alias{LocalPath}
\alias{LocalPath,FeatureList-method}
\alias{featurelist}
\title{FeatureList objects}
\usage{
FeatureList(..., names)

LocalPath(object)

\S4method{LocalPath}{FeatureList}(object)

Cached(object)

\S4method{Cached}{FeatureList}(object)
}
\arguments{
\item{...}{either the names of \code{data.frames}, \code{DataFrames}, or a
list of either class, optionally named}

\item{names}{a character vector of names to assign to each element}
}
\value{
\code{FeatureList}
}
\description{
FeatureList objects
}
\section{Methods}{


\describe{
 \item{\code{cache.features}}{Download uncached features}
}
}

\section{Accessors}{

\code{LocalPath}: Access LocalPath information from FeatureList object
\code{Cached}: Cached column from FeatureList object
}

