% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/featureOverlaps-methods.r
\name{featureOverlaps}
\alias{featureOverlaps}
\title{Find features overlapping GWAS markers}
\usage{
featureOverlaps(query, subject, maxgap = 0L, minoverlap = 1L,
  type = c("any", "start", "end", "within"), ...)
}
\arguments{
\item{query}{\code{GWAS} object}

\item{subject}{\code{FeatureList} object}

\item{maxgap}{
    Intervals with a separation of \code{maxgap} or less and a minimum
    of \code{minoverlap} overlapping positions, allowing for
    \code{maxgap}, are considered to be overlapping.  \code{maxgap}
    should be a scalar, non-negative, integer. \code{minoverlap}
    should be a scalar, positive integer.
  }

\item{minoverlap}{see documentation for \code{maxgap} argument}

\item{type}{
    By default, any overlap is accepted. By specifying the \code{type}
    parameter, one can select for specific types of overlap. The types
    correspond to operations in Allen's Interval Algebra (see
    references). If \code{type} is \code{start} or \code{end}, the
    intervals are required to have matching starts or ends,
    respectively. While this operation seems trivial, the naive
    implementation using \code{outer} would be much less
    efficient. Specifying \code{equal} as the type returns the
    intersection of the \code{start} and \code{end} matches. If
    \code{type} is \code{within}, the query interval must be wholly
    contained within the subject interval. Note that all matches must
    additionally satisfy the \code{minoverlap} constraint described above.

    The \code{maxgap} parameter has special meaning with the special
    overlap types. For \code{start}, \code{end}, and \code{equal},
    it specifies the maximum difference in the starts, ends or both,
    respectively. For \code{within}, it is the maximum amount by which
    the query may be wider than the subject.
  }

\item{...}{
    Further arguments to be passed to or from other methods:
    \itemize{
      \item \code{drop}: All methods accept the \code{drop} argument
            (\code{FALSE} by default). See \code{select} argument above
            for the details.
      \item \code{ignoreSelf}, \code{ignoreRedundant}: When \code{subject}
            is omitted, the \code{ignoreSelf} and \code{ignoreRedundant}
            arguments (both \code{FALSE} by default) are allowed.
            See \code{query} and \code{subject} arguments above for the
            details.
    }
  }
}
\value{
A list of \code{DataFrame}s, the structure of which mirrors that of
the supplied \code{FeatureList} object. Each element of the list is a
\code{DataFrame} with one logical vector for each feature, indicating
marker overlap.
}
\description{
Determine which GWAS markers overlap each \code{FeatureList} feature.
}

