% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotatedGWAS-methods.r
\name{features}
\alias{features}
\title{Extract list of features from AnnotatedGWAS object}
\usage{
features(object)
}
\arguments{
\item{object}{\code{AnnotatedGWAS} object}
}
\description{
Extract list of features from AnnotatedGWAS object
}

