% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GWAS-class.r
\docType{class}
\name{GWAS-class}
\alias{GWAS}
\alias{GWAS-class}
\title{GWAS-class}
\description{
The GWAS class is a container for representing the genetic markers used in a
genome-wide association study (GWAS), along with their corresponding summary
statistics.
}
\details{
The GWAS class is a \link[GenomicRanges]{GRanges} object with \code{marker}
and \code{pvalue} metadata columns. In addition, a GWAS object must also
contain a metadata column indicating the direction of the minor allele's
effect, which can be provided as one of the following:

\itemize{
 \item \code{zscore} z-scores
 \item \code{or} Odds ratios
 \item \code{beta} Beta values (i.e., the regression coefficient)
}
}
\section{Constructor}{


Use \code{\link{as.GWAS}} to construct a \code{GWAS} object from either a
\code{data.frame} or \code{GRanges} object
}

\section{Accessors}{


In the following code snippets, \code{x} is a \code{GWAS} object.

\describe{
 \item{}{
   \code{pvalue(x)} retrieve GWAS markers
 }
 \item{}{
   \code{pvalue(x)} retrieve marker pvalues
 }
 \item{}{
   \code{zscore(x)} retrieve marker z-scores
 }
}
}

