% Generated by roxygen2 (4.0.0): do not edit by hand
\name{load.feature}
\alias{load.feature}
\title{Load a local GRanges feature}
\usage{
load.feature(path)
}
\arguments{
  \item{path}{Path of feature to load}
}
\description{
\code{path} must point to an \code{R} object that contains a
\code{GRanges} feature
}

