% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/features.r
\name{local.features}
\alias{local.features}
\title{Retrieve information about local features}
\usage{
local.features(query = NULL, path)
}
\arguments{
\item{query}{a vector of strings to filter features; may also be a named list
of vectors to perform multiple queries for different categories of features}

\item{path}{Path of directory containing features}
}
\description{
Retrieve information about local features
}

