% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWAS-methods.r
\docType{methods}
\name{marker}
\alias{marker}
\alias{marker,GWAS-method}
\title{marker accessor and setter}
\usage{
marker(object)

\S4method{marker}{GWAS}(object)
}
\arguments{
\item{object}{an object to get or set the markers of}
}
\description{
Gets and sets an object's markers
}
\section{Methods (by class)}{
\itemize{
\item \code{GWAS}: Get and set a \code{GWAS} object's markers
}}

