\name{pull.features}
\alias{pull.features}
\title{Accessor for feature annotations stored in gwas object}
\usage{
pull.features(gwas)
}
\arguments{
  \item{gwas}{a \code{\link{GRanges}} object}
}
\value{
List of data.frames in which each slot corresponds to a
feature type
}
\description{
Only needed until feature annotations are stored
hierarchically by type
}

