% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWAS-methods.r
\docType{methods}
\name{pvalue}
\alias{pvalue}
\alias{pvalue,GWAS-method}
\title{pvalue accessor and setter}
\usage{
pvalue(object)

\S4method{pvalue}{GWAS}(object)
}
\arguments{
\item{object}{an object to get or set the pvalue of}
}
\description{
Gets and sets an object's pvalues
}
\section{Methods (by class)}{
\itemize{
\item \code{GWAS}: Get and set the pvalues of a \code{GWAS} object
}}

