\name{.validColClass}
\alias{.validColClass}
\title{Verify the presence and class of a data.frame variable}
\usage{
.validColClass(object, description, var.name, var.class)
}
\arguments{
  \item{object}{\code{data.frame} or a
  \code{\link[IRanges]{DataFrame}}}

  \item{description}{description of class being tested to
  print in error}

  \item{var.name}{column or slot name to check for}

  \item{var.class}{expected class of column or slot}
}
\description{
Verify the presence and class of a data.frame variable
}

