% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.r
\name{write-gwas}
\alias{write-gwas}
\alias{write_igv}
\title{Save GWAS and AnnotatedGWAS objects to an external file}
\usage{
write_igv(object, filename, ...)
}
\description{
Write GWAS objects to external files in various standard formats
}

