% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWAS-methods.r
\docType{methods}
\name{zscore}
\alias{zscore}
\alias{zscore,GWAS-method}
\title{zscore accessor and setter}
\usage{
zscore(object)

\S4method{zscore}{GWAS}(object)
}
\arguments{
\item{object}{an object to get or set the zscore of}
}
\description{
Gets and sets an object's zscores
}
\section{Methods (by class)}{
\itemize{
\item \code{GWAS}: Get and set a \code{GWAS} object's zscores
}}

