% Generated by roxygen2 (4.0.0): do not edit by hand
\name{annotate.gwas}
\alias{annotate.gwas}
\title{Annotate GWAS makers with genomic features}
\usage{
annotate.gwas(object, feature.list)
}
\arguments{
  \item{object}{\code{\link{GWAS}} object}

  \item{feature.list}{\code{FeatureList} object created
  with \code{\link{hub.features}} or
  \code{\link{local.features}}}
}
\value{
\code{\link{AnnotatedGWAS}} object
}
\description{
Annotate GWAS makers with genomic features
}
\examples{
query <- list(DNaseI = c("Dnasen", "broadPeak", "rep1"))
feature.list <- hub.features(query)
scz <- annotate.gwas(scz, feature.list)
}

