% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GWAS-class.r
\docType{methods}
\name{as.GWAS}
\alias{as.GWAS}
\alias{as.GWAS,GRanges-method}
\alias{as.GWAS,data.frame-method}
\title{Create a GWAS object}
\usage{
as.GWAS(object, genome, marker, chr, bp, pvalue, zscore, or, beta)

\S4method{as.GWAS}{data.frame}(object, genome, marker, chr, bp, pvalue, zscore,
  or, beta)

\S4method{as.GWAS}{GRanges}(object, genome, marker, pvalue, zscore, or, beta)
}
\arguments{
\item{object}{a \code{\link{data.frame}} or
\code{\link[GenomicRanges]{GRanges}} object containing the required fields
necessary to construct a \code{\link{GWAS}} object}

\item{genome}{genome abbreviation used by UCSC (e.g., hg19)}

\item{marker}{name of the column in \code{object} that contains the marker
(or SNP) identifiers}

\item{chr}{name of the column in \code{object} that contains the chromosome
names associated with each marker}

\item{bp}{name of the column in \code{object} that contains the genomic
position (or start position) associated with each marker}

\item{pvalue}{name of the column in \code{object} that contains the GWAS
association p-value for each marker}

\item{zscore}{name of the column in \code{object} that contains the GWAS z-scores
for each marker}

\item{or}{name of the column in \code{object} that contains the GWAS odds
ratio for each marker}

\item{beta}{name of the column in \code{object} that contains the GWAS beta
values (or regression coefficients) for each marker}
}
\description{
Create a \code{\link{GWAS}} object from a \code{\link{data.frame}} or a
\code{\link[GenomicRanges]{GRanges}} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Create a \code{\link{GWAS}} object from a
\code{\link{data.frame}}

\item \code{GRanges}: Create a \code{\link{GWAS}} object from a
\link[GenomicRanges]{GRanges}
object
}}

