\name{as.GWAS}
\alias{as.GWAS}
\title{Create a GWAS object}
\usage{
as.GWAS(object, marker, chr, bp, pvalue, or, beta)
}
\arguments{
  \item{snpid}{column label corresponding to SNP ID column}

  \item{chr}{column label corresponding to chromosome
  column}

  \item{pos}{}

  \item{pval}{}

  \item{or}{}
}
\description{
Create a GWAS object
}

