% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bingd-package.r
\docType{package}
\name{bingd}
\alias{bingd}
\alias{bingd-package}
\title{bingd: bayesian integration of genomic data}
\description{
bingd provides functions for conveniently integrating GWAS summary
statistics with various kinds of genomic features, including DNaseI
hypersensitivity data and evolutionary conservation scores
}
\author{
Aaron Wolen \email{aaron@wolen.com},
Mark Reimers \email{mreimers@vcu.edu}

Maintainer: Aaron Wolen \email{aaron@wolen.com}
}

