% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureList-methods.r
\name{cache.features}
\alias{cache.features}
\title{Download uncached features}
\usage{
cache.features(object, path)
}
\arguments{
\item{object}{\code{\link{GWAS}} object}

\item{path}{Path of directory containing features}
}
\value{
\code{FeatureList}
}
\description{
Download uncached features
}

