% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FeatureList-methods.r
\docType{methods}
\name{cache.features}
\alias{cache.features}
\alias{cache.features,FeatureList-method}
\title{Download uncached AnnotationHub features}
\usage{
cache.features(object, path)

\S4method{cache.features}{FeatureList}(object, path)
}
\arguments{
\item{object}{\code{\link{GWAS}} object}

\item{path}{Path of directory containing features}
}
\value{
\code{FeatureList}
}
\description{
Download uncached AnnotationHub features
}

