% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-conditionals.r
\docType{methods}
\name{calc.conditionals}
\alias{calc.conditionals}
\alias{calc.conditionals,AnnotatedGWAS-method}
\title{Calculate conditional probabilities for annotated features}
\usage{
calc.conditionals(object, risk.thresh = NULL, adjust = NULL,
  verbose = FALSE)

\S4method{calc.conditionals}{AnnotatedGWAS}(object, risk.thresh = NULL,
  adjust = NULL, verbose = FALSE)
}
\arguments{
\item{object}{\code{AnnotatedGWAS} object}

\item{risk.thresh}{the p-value threshold applied to identify ``risky'' markers}

\item{adjust}{optional parameter to adjust baseline and conditional
probabilities for linkage disequilibrium}

\item{verbose}{print out intermediate statistics calculated during the analysis}
}
\value{
data.frame containing probabilities of observing each combination of
annotated features.
}
\description{
Calculate conditional probabilities for annotated features
}

