\docType{methods}
\name{calc.enrich,AnnotatedGWAS-method}
\alias{calc.enrich,AnnotatedGWAS-method}
\title{Calculate enrichment of annotated features across thresholds}
\usage{
\S4method{calc.enrich}{AnnotatedGWAS}(object, stat, thresh.levels)
}
\description{
Calculate enrichment of annotated features across
thresholds
}

