\docType{methods}
\name{calc.enrich,GWAS-method}
\alias{calc.enrich,GWAS-method}
\title{Calculate enrichment of FeatureList features across thresholds}
\usage{
\S4method{calc.enrich}{GWAS}(object, feature.list, stat, thresh.levels)
}
\description{
Calculate enrichment of FeatureList features across
thresholds
}

