\name{consolidate}
\alias{consolidate}
\title{Consolidate each group of annotated features}
\usage{
consolidate(object)
}
\arguments{
  \item{\code{AnnotatedGWAS}}{object}
}
\value{
\code{AnnotatedGWAS} object
}
\description{
consolidate combines the features within each group to
produce a new single feature that comprises all of the
original individual features
}

