% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotatedGWAS-methods.r
\docType{methods}
\name{fcols}
\alias{fcols}
\alias{fcols,AnnotatedGWAS-method}
\title{Extract features from AnnotatedGWAS object}
\usage{
fcols(object)

\S4method{fcols}{AnnotatedGWAS}(object)
}
\arguments{
\item{object}{\code{AnnotatedGWAS} object}
}
\description{
Extract features from AnnotatedGWAS object
}

