% Generated by roxygen2 (4.0.0): do not edit by hand
\name{feature.labels}
\alias{feature.labels}
\title{Create pretty feature labels from the full RDataPaths}
\usage{
feature.labels(x)
}
\arguments{
  \item{x}{characer vector of full RDataPaths}
}
\description{
Basically just strips off the extension from the basename
}

