% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureList-class.r
\name{FeatureList}
\alias{FeatureList}
\title{FeatureList objects}
\usage{
FeatureList(..., names)
}
\arguments{
\item{...}{either the names of \code{data.frames}, \code{DataFrames}, or a
list of either class, optionally named}

\item{names}{a character vector of names to assign to each element}
}
\value{
\code{FeatureList}
}
\description{
FeatureList objects
}

