\name{features}
\alias{features}
\title{Accessor for feature annotations}
\usage{
features(object)
}
\arguments{
  \item{gwas}{GWAS GRanges object}
}
\value{
List of data.frames in which each slot corresponds to a
feature type
}
\description{
Only needed until feature annotations are stored
hierarchically by type
}

