% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AnnotatedGWAS-methods.r
\docType{methods}
\name{features}
\alias{features}
\alias{features,AnnotatedGWAS-method}
\title{Extract list of features from AnnotatedGWAS object}
\usage{
features(object)

\S4method{features}{AnnotatedGWAS}(object)
}
\arguments{
\item{object}{\code{AnnotatedGWAS} object}
}
\description{
Extract list of features from AnnotatedGWAS object
}

