% Generated by roxygen2 (4.0.0): do not edit by hand
\name{filter.features}
\alias{filter.features}
\title{Filter FeatureList object based on search terms}
\usage{
filter.features(object, query)
}
\arguments{
  \item{query}{a vector of strings to filter features; may
  also be a named list of vectors to perform multiple
  queries for different categories of features}

  \item{object}{a \code{\link{FeatureList}} object
  comprising one or more DataFrames that containing, at
  minimum, \code{Title} \code{LocalPath}, and \code{Cached}
  columns}
}
\description{
Search terms can be grouped together using a list of queries.
}

