% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GWAS-class.r
\docType{class}
\name{GWAS-class}
\alias{GWAS}
\alias{GWAS-class}
\alias{marker}
\alias{pvalue}
\alias{zscore}
\title{GWAS-class}
\description{
The GWAS class is a container for representing the genetic markers used in a
genome-wide associate study (GWAS), along with their corresponding summary
statistics.
}
\details{
The GWAS class is a \link[GenomicRanges]{GRanges} with the following required
metadata columns:

\describe{
 \item{\code{marker}}{Marker or SNP identifiers}
 \item{\code{pvalue}}{Marker association p-values}
}

A GWAS object must also contain a metadata column indicating the direction of
the minor allele's effect, which can be provided as one of the following:

\describe{
 \item{\code{zscore}}{z-scores}
 \item{\code{or}}{Odds ratios}
 \item{\code{beta}}{Beta values (i.e., the regression coefficient)}
}
}
\section{Accessors}{


In the following code snippets, \code{x} is a \code{GWAS} object.

\describe{
 \item{}{
   \code{pvalue(x)} retrieve GWAS markers
 }
 \item{}{
   \code{pvalue(x)} retrieve marker pvalues
 }
 \item{}{
   \code{zscore(x)} retrieve marker z-scores
 }
}
}

