% Generated by roxygen2 (4.0.0): do not edit by hand
\name{local.features}
\alias{local.features}
\title{Retrieve information about local features}
\usage{
local.features(query = NULL, path)
}
\arguments{
  \item{path}{Path of directory containing features}

  \item{query}{a vector of strings to filter features; may
  also be a named list of vectors to perform multiple
  queries for different categories of features}
}
\description{
Retrieve information about local features
}

