% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureList-methods.r
\name{LocalPath}
\alias{LocalPath}
\title{Access LocalPath information from FeatureList object}
\usage{
LocalPath(object)
}
\arguments{
\item{object}{\code{FeatureList} object}
}
\description{
Access LocalPath information from FeatureList object
}

