% Generated by roxygen2 (4.0.0): do not edit by hand
\name{serial.enrich}
\alias{serial.enrich}
\title{Calculate the enrichment of a particular discrete feature across thresholds}
\usage{
serial.enrich(feature, stat, thresh.levels)
}
\arguments{
  \item{feature}{logical or character vector}

  \item{stat}{numeric vector containing statistic to which
  thresholds will be applied for each enrichment
  calculation}

  \item{thresh.levels}{a numeric vector containing the
  various thresholds at which \code{feature} enrichment is
  calculated}
}
\value{
data.frame containing `enrichment`
}
\description{
Calculate the enrichment of a particular discrete feature across thresholds
}

